requirements.txt
feedparser
python-dotenv